After a while, a huge commit, that mainly introduce code for window managment. That means two new modules (based on xlib), with code from dminiwm, xkev and xcut for a start. But probably if time permits much code will change as it is really just a first functional prototype, but which for sure it contains bugs (though not even a single crash for a couple of days).
So because of this feature, lets call it 0.9.3 and lets dedicated (first to Lue Reed?
(I'm not sure about the spelling and there is no Internet around to check), as I just
heard "vicius" on the radio), but especially to my hero (though there arn't heroes
anymore) Iggy Pop. I'm just another lucky earth passenger.