Rline: handle the case of ../../.. type paths. Important:
debuging led me to the root of the cause, which is the handling of such
paths in path_dirname () S-Lang intrinsic function, which in those
cases returns an empty string.

by trying to overcome this behavior realized that everything seems better
when relative paths are translated to absolute paths and indeed this makes
sense. This became the default interpetation with the only exception so
far the `ln' command

also, commited and since established code changed in a way that it might
introduce unexpected results, a couple of printing debugging messages,
with the hope it will catch errors in the normal workflow