media: abort an endless loop while there is no other song in the playlist and mplayer (probably) don't send any data to the fifo
also, don't use pcre routines to find a lyric (a matched lyric now is just
the name of the song minus the extension[s]). But this also means that
we loose the ability to have a lyric name with different case than the song
name.
Thanks to Dimitris for both remarks.