Mainly adjustments for pointer/indexes calculations on operations when lines > available screen size
The general idea - on almost any operation - is to check the pointer
position, if it is at the first or last available cell (sometimes and
where is appropriate the next or previous respectively) and to operate
depending of, if the line length is longer than the available screen
size and the pointer is at a state where the first visible char is
the actual first of the current line. That's the design which came out,
and is not easy, as many conditions has to hand checked, but if its
done once and its not wrong cannot be wrong then.
The hard part is that according with the initial thoughts - that the
editor had to support an optional indent and maximum length - this is
hard to get it right. The code there has flaws as it has to compare
the available cells number. For now everything is based on the initial
(fixed) size. The first checks are easy anyway but what if the screen
changes? Nothing in the code so far is prepared for a sigwinch.