__F: this fixes another weakness in that code, specifically the case when there are no declared arguments and the first statement discards the returned value from a function that returns a value, e.g.,
		() = fprintf (...

it also handles the "()=" case, though this is a violation in the
coding style, which simple separate with a space (almost)
any token - with the exceptions:
a. when getting the index of an array, and use (simple arithmetic)
e.g., "[idx+1]"
b. when functions declared like __fun__ and arguments declared
the same way, e.g., __fun__(__arg__, ...), this can become
sometimes an amazing ascci art (and still tokens are
clearly separated)

the reason: (optical) clarity
cons: add another byte (many actually) but the value for this
waste is incalculable and quite precious to not to buy