Commited on __dev a prototype which implements the following consepts:
Every function can has its own private environment, where the state
is controlled only within the function environment, in which nested
functions can also be defined.

That means one function can also be the program.

Such an environment, where functions share same properties and are
related to each other, is called a class, which can be called as
standalone, or as a new copy (where the environment is initialized
to the default state), or as a subclass of another class, which
also (like functions) can have unlimited nested levels

Class functions have access to self, which is the executed
function structure, and to this, which is the current state

Commited also a sample Piple Class and the two required classes
Var and Type.

The drawback to this implementation is that arguments are strings
that evaluated on the fly, and there isn't another way for these
multiline strings other than using backquotes, where they should be
doubled on nested declarations.