Array: added __extend () a practical but total unsafe function that do not do any kind of check; but if it is so why then?
this is just to declare the intention that the human energy should be
better spend on the idea[s] and to the actual implementation of them,
rather to code expression which should be feel natural and compatible
with the way of thinking; (in that case the natural flow is):
  ar =  Array.New ([1, 2]); ar.extend with (elem, ...)
there is no real need to specify the starting position to place the
elements, because the usual is to append the elements at the end;
if another position is desired, then an attribute can be declared at
the function call level:
  ar.extend (elem, ...;prepend)  or  ar.extend (elem, ...;at = index)

in S-Lang is easy to implement this object style like, by simply defining
a structure (with a get method and a field to hold the array value) dressed
with varius methods. but this makes the code uneccecary complex when a simple:
  ar = [ar, elem, ...] will do
but the practicallity remains and holds quite of a strong value

in this implementation however there is an established style, and thats why
the array is passed as a reference; as for the sanity checks? the
responsibility belongs to the coder (who is the one who needs protection
by obvious mistakes); quite a contradiction but this also called balance.

some people however should be code the low level details and as long this
condition is met, the system works (it is believed now that there are those
that have this will)