___: introduce a lot of code to handle library dependencies
This started because of libcurl (which is a prerequisite), that links
against a lot of libraries, that might not match the default flags.

However its good idea to expand to a more accurate building system,
though the modules trying to be minimal, and|or explicitly ask for
a specific library which in that case sets the necessary flags.
In any case this is a start for a building system and objdump(1) is
installed by default with binutils