Ved: added __vhandle_word__ () that is called on normal mode to handle operations on current word, and its triggered with "W".
For now only one function is defined which is to translate current word. And for
this a translate() method added to Lingual subclass. This looks to
Env->USER_DATA_PATH + "/dict" for a dictionary based on the qualifier dict, or
a "dict.txt" file by default. It returns an array which are the lines from the
dictionary that match <word>

(also) renamed Spell to Lingual