Rline: trying to fix and old drawing bug, on the function that only cares about filenames. The trueth is that this function just exposed a weakness on the machine, which is rather fragile. Anyway the misbehavior came from a passing qualifier that came from a long path to be easy to trace it. Remind me, that except the incredible power that qualifiers give, is that its better when they go too deep, to reset them in one point and recreate them. And indeed looks wise to give unique names to qualifiers that makes sense and are understandable.
(also) added define_key() intrinsic function to map keycodes to chars in the
Input layer. As a first thing that did is to create a hellenic map which works
perfectly. The only thing is that I'm not sure yet, if it's the right thing, to
call it at every re-initialization of the tty.