Rline: selection function: try to give the focus to the correct item, on some corner cases
reminds me, that this function started as a quick widget (at the very early steps)
and ended up as the most precious function as it does all these selections which
is a joy to use. But at the same time is the most complicated. The calculations
were always a puzzle. It didn't get much of love and the trueth is that, since
it is working i do not want to enter there :-)