This is the last commit of this development circle. It's been a little more than two years, with 584 commits, of which some of them could easily be regarded as more than one.
This work is dedicated to the ones I feel more connected to, that is:

The Thessaly tribe, the Nepali tribe, the Apache tribe and the mountain Mexican
and Japanese villages, but and to all the farmers and to shephers of this world.

To all the self educated people.

And to all that has been accepted their nature, but choosed to continue living
by making their own life part of their art, that simply means:

Da Da.

THIS IS THE END